services/backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pillow==9.4.0
pandas
matplotlib
seaborn
numpy==1.24.4
torch==2.2.2+cpu
torchvision==0.17.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
